Use O_CLOEXEC and O_NONBLOCK from the os module
Instead of hard-coding them. It turns out they are
architecture-specific.

What we really want are `IN_CLOEXEC` and `IN_NONBLOCK`, but they appear
to be identical to the constants in the `os` module, the `IN_*`
constants are not available through the interpreter, and we don't want
to have to parse C header files to get them.

Addresses an issue raised in the discussion thread for #20