Fix readthedocs URL in documentation
While reading the README, I noticed that when visiting the URL
http://inotify_simple.readthedocs.org one gets a `Bad Request (400)`
error.  I made a guess that the underscore would might be a hyphen on
the readthedocs site and it turned out I was right.  This commit fixes
the issue and thus allows readers of the docs and the README to find the
readthedocs page successfully.